Tapping on a custom status in the message list opens the users profile card
